+++ src/openzaak/config/templates/config/config_external.html
@@ -9,9 +9,11 @@
 </div>
 
 <div class="container">
-    {{ formdata|json_script:"formdata" }}
     {{ auth_types|json_script:"auth-type-choices" }}
 {#    {{ nlx_choices|json_script:"nlx-choices" }}#}
+
+    {{ formset_config|json_script:"formset-config" }}
+    {{ formdata|json_script:"formdata" }}
     <form method="post" enctype="multipart/form-data">
         {% csrf_token %}
 
@@ -19,32 +21,32 @@
             {# empty on purpose - react managed #}
         </div>
 
-        {% load_widgets form='snippets/forms.html' %}
-
-        {{ formset.management_form }}
-        {{ formset.non_form_errors }}
-
-        <div class="form-group row text-center">
-            <div class="col"><strong>Label</strong></div>
-            <div class="col"><strong>API type</strong></div>
-            <div class="col-sm-3"><strong>Url</strong></div>
-            <div class="col"><strong>NLX</strong></div>
-            <div class="col"><strong>Authorization</strong></div>
-        </div>
-
-        {% for form in formset %}
-            {{ form.non_field_errors }}
-            {{ form.id }}
-
-            <div class="row">
-                <div class="form-group col">{% form_field form.label %}</div>
-                <div class="form-group col">{% form_field form.api_type %}</div>
-                <div class="form-group col-sm-3">{% form_field form.api_root widget='form:URLInput' %}</div>
-                <div class="form-group col">{% form_field form.nlx %}</div>
-                <div class="form-group col">{% form_field form.auth_type %}</div>
-            </div>
-
-        {% endfor %}
+{#        {% load_widgets form='snippets/forms.html' %}#}
+{##}
+{#        {{ formset.management_form }}#}
+{#        {{ formset.non_form_errors }}#}
+{##}
+{#        <div class="form-group row text-center">#}
+{#            <div class="col"><strong>Label</strong></div>#}
+{#            <div class="col"><strong>API type</strong></div>#}
+{#            <div class="col-sm-3"><strong>Url</strong></div>#}
+{#            <div class="col"><strong>NLX</strong></div>#}
+{#            <div class="col"><strong>Authorization</strong></div>#}
+{#        </div>#}
+{##}
+{#        {% for form in formset %}#}
+{#            {{ form.non_field_errors }}#}
+{#            {{ form.id }}#}
+{##}
+{#            <div class="row">#}
+{#                <div class="form-group col">{% form_field form.label %}</div>#}
+{#                <div class="form-group col">{% form_field form.api_type %}</div>#}
+{#                <div class="form-group col-sm-3">{% form_field form.api_root widget='form:URLInput' %}</div>#}
+{#                <div class="form-group col">{% form_field form.nlx %}</div>#}
+{#                <div class="form-group col">{% form_field form.auth_type %}</div>#}
+{#            </div>#}
+{##}
+{#        {% endfor %}#}
 
 
         <div class="btn-toolbar justify-content-between py-sm-3">
