+++ src/openzaak/js/components/external-services/external-formset.js
@@ -1,9 +1,10 @@
 import React from "react";
 import {ExternalForm} from "./external-form";
+import {ManagementForm} from "../../forms/management-form";
 
 
 function ExternalFormSet(props) {
-    const formData = props.formData;
+    const { config, formData } = props;
 
     // set up the existing forms
     const forms = formData.map(
@@ -11,7 +12,15 @@ function ExternalFormSet(props) {
     );
 
     return (
-        <div>
+        <>
+            <ManagementForm
+                prefix={config.prefix}
+                initial_forms={config.INITIAL_FORMS}
+                total_forms={ formData.length }
+                min_num_forms={config.MIN_NUM_FORMS}
+                max_num_forms={config.MAX_NUM_FORMS}
+            />
+
             <div className="form-group row text-center">
                 <div className="col"><strong>Label</strong></div>
                 <div className="col"><strong>API type</strong></div>
@@ -21,7 +30,7 @@ function ExternalFormSet(props) {
             </div>
 
             {forms}
-        </div>
+        </>
     );
 
 }
